The `replace` in conf/modules.config only affects the last extension
This means we need to use "baseName" rather than "simpleName". It makes a
difference if there are multiple dots in the file name.

For instance, "a.b.bedpe" will be published as "a.b.genome" by Nextflow, but
the "simpleName" transformation would make "a.genome".

`.getBaseName()` is the same as `.baseName`.