Get the BUSCO ODB name from the BUSCO definition file rather than GoaT
The issue with GoaT is that it can only handle assemblies that BTK results and
there are issues retrieving older assemblies.

Since the BUSCO ODBs are related to the taxon, not the assembly, start over
from the BUSCO ODB <-> taxon mapping file, query the NCBI to get the list of
all ancestors of that taxon (which works regardless of the assemblies) and get
the most recent one.